correct folder name and README title
