+++ exercise_solutions/ex2/main.go
@@ -1,15 +1,40 @@
 package main
 
-import "fmt"
+import (
+	"fmt"
+	"io"
+	"log"
+	"os"
+)
 
-func prefixer(prefix string) func(string) string {
-	return func(body string) string {
-		return prefix + " " + body
+func fileLen(fileName string) (int, error) {
+	f, err := os.Open(fileName)
+	if err != nil {
+		return 0, err
 	}
+	defer f.Close()
+	data := make([]byte, 2048)
+	total := 0
+	for {
+		count, err := f.Read(data)
+		total += count
+		if err != nil {
+			if err != io.EOF {
+				return 0, err
+			}
+			break
+		}
+	}
+	return total, nil
 }
 
 func main() {
-	helloPrefix := prefixer("Hello")
-	fmt.Println(helloPrefix("Bob"))   // should print Hello Bob
-	fmt.Println(helloPrefix("Maria")) // should print Hello Maria
+	if len(os.Args) < 2 {
+		return
+	}
+	count, err := fileLen(os.Args[1])
+	if err != nil {
+		log.Fatal(err)
+	}
+	fmt.Println(count)
 }
